Inventory_Managment_System/requirements.txt
fpdf2==2.7.8
PyInstaller==6.3.0
pillow==10.2.0
